Renamed MeterReader components
