+++ src/components/features/data-tables/meter-reader-entry/MeterReaderEntryDataTable.tsx
@@ -1,27 +1,25 @@
 "use client";
 
 import { FunctionComponent, Suspense } from "react";
-import { useMeterReaderEntryColumns } from "./MeterReaderEntryColumns";
-import { useMeterReadersStore } from "@mr/components/stores/useMeterReadersStore";
 import { DataTable } from "@mr/components/ui/data-table/data-table";
+import { useMeterReaderEntryColumns } from "./MeterReaderEntryColumns";
+import { useSchedulesStore } from "@mr/components/stores/useSchedulesStore";
+import { MeterReaderWithZonebooks } from "@mr/lib/types/personnel";
 
-export const MeterReaderEntryDataTable: FunctionComponent = () => {
-  const meterReaders = useMeterReadersStore((state) => state.meterReaders);
+type MeterReaderDataTableProps = {
+  meterReaders: MeterReaderWithZonebooks[];
+};
 
-  // const { data: meterReaders } = useQuery({
-  //   queryKey: ["get-all-meter-readers"],
-  //   queryFn: async () => {
-  //     const res = await axios.get(`${process.env.NEXT_PUBLIC_MR_BE}/personnel?status=assigned`);
+export const MeterReaderEntryDataTable: FunctionComponent<MeterReaderDataTableProps> = ({ meterReaders }) => {
+  // const [meterReaders, setMeterReaders] = useState<MeterReader[]>([]);
 
-  //     return res.data;
-  //   },
-  // });
+  const selectedScheduleEntry = useSchedulesStore((state) => state.selectedScheduleEntry);
 
-  const meterReaderEntryColumns = useMeterReaderEntryColumns(meterReaders);
+  const meterReaderEntryColumns = useMeterReaderEntryColumns(selectedScheduleEntry?.meterReaders);
 
   return (
     <Suspense fallback={<p>Loading...</p>}>
-      <DataTable data={meterReaders ?? []} columns={meterReaderEntryColumns} />
+      <DataTable data={meterReaders ? meterReaders : []} columns={meterReaderEntryColumns} />
     </Suspense>
   );
 };
