+++ src/components/features/data-tables/meter-reader-entry/MeterReaderEntryRowActions.tsx
@@ -1,64 +1,68 @@
+/* eslint-disable @typescript-eslint/no-non-null-asserted-optional-chain */
 "use client";
 
+import { useSchedulesStore } from "@mr/components/stores/useSchedulesStore";
 import { Button } from "@mr/components/ui/Button";
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuSeparator,
-  DropdownMenuTrigger,
-} from "@mr/components/ui/DropdownMenu";
-import { MoreHorizontal, MoreVertical } from "lucide-react";
-import { FunctionComponent, useState } from "react";
-import { EditMeterReaderDialog } from "../../meter-readers/EditMeterReaderDialog";
-import { useMeterReadersStore } from "@mr/components/stores/useMeterReadersStore";
-import { MeterReader } from "@mr/lib/types/personnel";
-import { DeleteMeterReaderDialog } from "../../meter-readers/DeleteMeterReaderDialog";
+import { MeterReaderWithZonebooks } from "@mr/lib/types/personnel";
+import { CircleXIcon, MapPinnedIcon } from "lucide-react";
+import { FunctionComponent } from "react";
+import { ScheduleEntryZonebookSelector } from "../../scheduler/entry/ScheduleEntryZonebookSelector";
 
 type MeterReaderEntryRowActionsProps = {
-  meterReader: MeterReader;
+  meterReader: MeterReaderWithZonebooks;
 };
+
+// zone 8-63
 
 export const MeterReaderEntryRowActions: FunctionComponent<MeterReaderEntryRowActionsProps> = ({
   meterReader,
 }) => {
-  const [editMeterReaderDialogIsOpen, setEditMeterReaderDialogIsOpen] = useState<boolean>(false);
+  const selectedScheduleEntry = useSchedulesStore((state) => state.selectedScheduleEntry);
+  const setSelectedScheduleEntry = useSchedulesStore((state) => state.setSelectedScheduleEntry);
+  const setSelectedMeterReader = useSchedulesStore((state) => state.setSelectedMeterReader);
+  const setEntryZonebookSelectorIsOpen = useSchedulesStore((state) => state.setEntryZonebookSelectorIsOpen);
+  const setMeterReaderZonebooks = useSchedulesStore((state) => state.setMeterReaderZonebooks);
 
-  const setSelectedMeterReader = useMeterReadersStore((state) => state.setSelectedMeterReader);
+  const openZonebookSelector = (meterReader: MeterReaderWithZonebooks) => {
+    setSelectedMeterReader(meterReader);
+    setEntryZonebookSelectorIsOpen(true);
+    setMeterReaderZonebooks(meterReader.zonebooks ?? []);
+  };
+
+  const removeMeterReader = (companyId: string) => {
+    const temporaryMeterReaders = [...selectedScheduleEntry!.meterReaders!];
+
+    setSelectedScheduleEntry({
+      ...selectedScheduleEntry,
+      readingDate: selectedScheduleEntry?.readingDate!,
+      disconnectionDate: selectedScheduleEntry?.disconnectionDate!,
+      dueDate: selectedScheduleEntry?.dueDate!,
+      meterReaders: temporaryMeterReaders.filter((mr) => mr.companyId !== companyId),
+    });
+  };
 
   return (
-    <DropdownMenu>
-      <DropdownMenuTrigger asChild>
-        <Button variant="ghost" className="data-[state=open]:bg-muted flex h-8 w-8 p-0">
-          <div className="hidden sm:hidden md:block lg:block">
-            <MoreHorizontal />
-          </div>
-          <div className="block sm:block md:hidden lg:hidden">
-            <MoreVertical />
-          </div>
-          <span className="sr-only">Open menu</span>
-        </Button>
-      </DropdownMenuTrigger>
-      <DropdownMenuContent align="end" className="w-[200px]">
-        <DropdownMenuItem
-          onClick={() => {
-            setSelectedMeterReader(meterReader);
-          }}
-          className="cursor-pointer"
-          asChild
-        >
-          {/* Update personnel */}
-          <EditMeterReaderDialog
-            editMeterReaderDialogIsOpen={editMeterReaderDialogIsOpen}
-            setEditMeterReaderDialogIsOpen={setEditMeterReaderDialogIsOpen}
-            selectedMeterReader={meterReader}
-          />
-        </DropdownMenuItem>
-        <DropdownMenuSeparator />
-        <DropdownMenuItem asChild>
-          <DeleteMeterReaderDialog selectedMeterReader={meterReader} />
-        </DropdownMenuItem>
-      </DropdownMenuContent>
-    </DropdownMenu>
+    <>
+      <ScheduleEntryZonebookSelector isLoading={false} />
+      <div className="flex grid-cols-2 gap-2">
+        <div className="col-span-1">
+          <Button className="w-full px-2" size="sm" onClick={() => openZonebookSelector(meterReader)}>
+            <MapPinnedIcon className="size-2 sm:size-4 lg:size-4 dark:text-white" />
+            <span className="hidden text-xs sm:hidden md:hidden lg:block dark:text-white"> Zonebooks</span>
+          </Button>
+        </div>
+        <div className="col-span-1">
+          <Button
+            variant="destructive"
+            className="w-full px-2"
+            size="sm"
+            onClick={() => removeMeterReader(meterReader.companyId)}
+          >
+            <CircleXIcon className="size-2 fill-red-600 text-white sm:size-4 lg:size-4" />
+            <span className="hidden text-xs sm:hidden md:hidden lg:block">Remove</span>
+          </Button>
+        </div>
+      </div>
+    </>
   );
 };
