+++ src/components/features/data-tables/meter-readers/MeterReaderRowActions.tsx
@@ -1,66 +1,62 @@
-/* eslint-disable @typescript-eslint/no-non-null-asserted-optional-chain */
 "use client";
 
-import { useSchedulesStore } from "@mr/components/stores/useSchedulesStore";
 import { Button } from "@mr/components/ui/Button";
-import { MeterReaderWithZonebooks } from "@mr/lib/types/personnel";
-import { CircleXIcon, MapPinnedIcon } from "lucide-react";
-import { FunctionComponent } from "react";
-import { ScheduleEntryZonebookSelector } from "../../scheduler/entry/ScheduleEntryZonebookSelector";
+import {
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuSeparator,
+  DropdownMenuTrigger,
+} from "@mr/components/ui/DropdownMenu";
+import { MoreHorizontal, MoreVertical } from "lucide-react";
+import { FunctionComponent, useState } from "react";
+import { EditMeterReaderDialog } from "../../meter-readers/EditMeterReaderDialog";
+import { useMeterReadersStore } from "@mr/components/stores/useMeterReadersStore";
+import { MeterReader } from "@mr/lib/types/personnel";
+import { DeleteMeterReaderDialog } from "../../meter-readers/DeleteMeterReaderDialog";
 
-type PersonnelRowActionsProps = {
-  meterReader: MeterReaderWithZonebooks;
+type MeterReaderRowActionsProps = {
+  meterReader: MeterReader;
 };
 
-// zone 8-63
+export const MeterReaderRowActions: FunctionComponent<MeterReaderRowActionsProps> = ({ meterReader }) => {
+  const [editMeterReaderDialogIsOpen, setEditMeterReaderDialogIsOpen] = useState<boolean>(false);
 
-export const MeterReaderRowActions: FunctionComponent<PersonnelRowActionsProps> = ({ meterReader }) => {
-  const selectedScheduleEntry = useSchedulesStore((state) => state.selectedScheduleEntry);
-  const setSelectedScheduleEntry = useSchedulesStore((state) => state.setSelectedScheduleEntry);
-  const setSelectedMeterReader = useSchedulesStore((state) => state.setSelectedMeterReader);
-  const setEntryZonebookSelectorIsOpen = useSchedulesStore((state) => state.setEntryZonebookSelectorIsOpen);
-  const setMeterReaderZonebooks = useSchedulesStore((state) => state.setMeterReaderZonebooks);
-
-  const openZonebookSelector = (meterReader: MeterReaderWithZonebooks) => {
-    setSelectedMeterReader(meterReader);
-    setEntryZonebookSelectorIsOpen(true);
-    setMeterReaderZonebooks(meterReader.zonebooks ?? []);
-  };
-
-  const removeMeterReader = (companyId: string) => {
-    const temporaryMeterReaders = [...selectedScheduleEntry!.meterReaders!];
-
-    setSelectedScheduleEntry({
-      ...selectedScheduleEntry,
-      readingDate: selectedScheduleEntry?.readingDate!,
-      disconnectionDate: selectedScheduleEntry?.disconnectionDate!,
-      dueDate: selectedScheduleEntry?.dueDate!,
-      meterReaders: temporaryMeterReaders.filter((mr) => mr.companyId !== companyId),
-    });
-  };
+  const setSelectedMeterReader = useMeterReadersStore((state) => state.setSelectedMeterReader);
 
   return (
-    <>
-      <ScheduleEntryZonebookSelector isLoading={false} />
-      <div className="flex grid-cols-2 gap-2">
-        <div className="col-span-1">
-          <Button className="w-full px-2" size="sm" onClick={() => openZonebookSelector(meterReader)}>
-            <MapPinnedIcon className="size-2 sm:size-4 lg:size-4 dark:text-white" />
-            <span className="hidden text-xs sm:hidden md:hidden lg:block dark:text-white"> Zonebooks</span>
-          </Button>
-        </div>
-        <div className="col-span-1">
-          <Button
-            variant="destructive"
-            className="w-full px-2"
-            size="sm"
-            onClick={() => removeMeterReader(meterReader.companyId)}
-          >
-            <CircleXIcon className="size-2 fill-red-600 text-white sm:size-4 lg:size-4" />
-            <span className="hidden text-xs sm:hidden md:hidden lg:block">Remove</span>
-          </Button>
-        </div>
-      </div>
-    </>
+    <DropdownMenu>
+      <DropdownMenuTrigger asChild>
+        <Button variant="ghost" className="data-[state=open]:bg-muted flex h-8 w-8 p-0">
+          <div className="hidden sm:hidden md:block lg:block">
+            <MoreHorizontal />
+          </div>
+          <div className="block sm:block md:hidden lg:hidden">
+            <MoreVertical />
+          </div>
+          <span className="sr-only">Open menu</span>
+        </Button>
+      </DropdownMenuTrigger>
+      <DropdownMenuContent align="end" className="w-[200px]">
+        <DropdownMenuItem
+          onClick={() => {
+            setSelectedMeterReader(meterReader);
+          }}
+          className="cursor-pointer"
+          asChild
+        >
+          {/* Update personnel */}
+          <EditMeterReaderDialog
+            editMeterReaderDialogIsOpen={editMeterReaderDialogIsOpen}
+            setEditMeterReaderDialogIsOpen={setEditMeterReaderDialogIsOpen}
+            selectedMeterReader={meterReader}
+          />
+        </DropdownMenuItem>
+        <DropdownMenuSeparator />
+        <DropdownMenuItem asChild>
+          <DeleteMeterReaderDialog selectedMeterReader={meterReader} />
+        </DropdownMenuItem>
+      </DropdownMenuContent>
+    </DropdownMenu>
   );
 };
