+++ tsconfig.json
@@ -19,7 +19,8 @@
       }
     ],
     "paths": {
-      "@mr/*": ["./src/*"]
+      "@mr/*": ["./src/*"],
+      "@images/*": ["./public/images/*"]
     }
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
